fix(auth): now properly fixed
The `SameSite` attribute has several values, we had it set at `Strict`:

- With `Strict`, the browser only sends the cookie with requests from
  the cookie's origin site.

That doesn't sound right, this means opening the website will **not**
send the cookie, and we need it to. Let's continue:

- `Lax` is similar, except the browser also sends the cookie when the
  user navigates to the cookie's origin site (even if the user is coming
  from a different site).

This sounds perfect! This behaviour is exactly what we wanted.

That being said, I should have read the documentation, I thought the
difference was for subdomains, but that's not the case.

I also refactored things a little, cleaning up the auth system.

For the record, `useSession` internally calls `getSession`.