refactor(auth): switch to renderless callback
`updateSession` has to be imported because Nuxt doesn't auto-inject
`h3`'s functions into `pages/` (despite typings saying otherwise),
however, it seems properly handled as the client code doesn't error,
since the page is SSR'd and Nuxt has access to it.

This solution is a bit hacky, but it achieves full renderless on
successful auth, the user is redirected without even seeing
`/auth/callback`, and the response time is increased as it skips two
roundtrips (one for the auth plugin and another for the callback).

I tried a lot of other solutions, including defining the same `key` for
both `useFetch` and even doing `await execute({ dedupe: true })`,
however, none of the solutions I could come up with worked. This is the
only one that did.

Now with this patch, we go 1 `<ClientOnly>` down, only one left now.